chore: move ConnectWalletMessage to molecules folder
